update(Components): Update Movable and Transform structs to use float instead of double

+++ Flakkari/Engine/EntityComponentSystem/Components/2D/Movable.hpp
@@ -16,13 +16,13 @@ namespace Flakkari::Engine::ECS::Components::_2D {
 PACKED_START
 
 struct Movable {
-    Math::Vector2d velocity; // pixels / second
-    double angularVelocity; // degrees / second
-    Math::Vector2d acceleration; // pixels / second^2
-    double angularAcceleration; // degrees / second^2
+    Math::Vector2f velocity; // pixels / second
+    float angularVelocity; // degrees / second
+    Math::Vector2f acceleration; // pixels / second^2
+    float angularAcceleration; // degrees / second^2
 
     Movable() : velocity(0, 0), acceleration(0, 0) {};
-    Movable(const Math::Vector2d &velocity, const Math::Vector2d &acceleration) : velocity(velocity), acceleration(acceleration) {};
+    Movable(const Math::Vector2f &velocity, const Math::Vector2f &acceleration) : velocity(velocity), acceleration(acceleration) {};
     Movable(const Movable &other) : velocity(other.velocity), acceleration(other.acceleration) {};
 
     std::size_t size() const {
